06_rag_fusion/requirements.txt
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0
langchain-community>=0.1.0
openai>=1.0.0
qdrant-client>=1.14.0
python-dotenv>=1.1.0
numpy>=1.23.0
